admin/requirements.txt
quart==0.19.4
jinja2==3.1.3
aiohttp==3.9.1
python-dotenv==1.0.0
